+++ CHANGELOG.md
@@ -130,11 +130,13 @@ The callback function provided as parameter to the format function is now called
 * Fix multiline if-statements `multiline-if-else` ([#828](https://github.com/pinterest/ktlint/issues/828))
 * Prevent class cast exception on ".editorconfig" property `ktlint_code_style`  ([#1559](https://github.com/pinterest/ktlint/issues/1559))
 * Handle trailing comma in enums `trailing-comma` ([#1542](https://github.com/pinterest/ktlint/pull/1542))
+* Allow EOL comment after annotation ([#1539](https://github.com/pinterest/ktlint/issues/1539))
 
 ### Changed
 
 * Print an error message and return with non-zero exit code when no files are found that match with the globs ([#629](https://github.com/pinterest/ktlint/issue/629)).
 * Invoke callback on `format` function for all errors including errors that are autocorrected ([#1491](https://github.com/pinterest/ktlint/issues/1491))
+* Improve rule `annotation` ([#1574](https://github.com/pinterest/ktlint/pull/1574))
 
 ### Removed